Fix for update failure - sets user permissions and allows update from omni

+++ spacepy/toolbox.py
@@ -6,7 +6,7 @@ Toolbox of various functions and generic utilities.
 """
 from __future__ import division
 from spacepy import help
-__version__ = "$Revision: 1.11 $, $Date: 2010/05/26 21:31:18 $"
+__version__ = "$Revision: 1.12 $, $Date: 2010/05/26 22:00:01 $"
 __author__ = 'S. Morley and J. Koller'
 
 
@@ -447,15 +447,15 @@ def printfig(fignum, saveonly=False, pngonly=False, clean=False):
    return
 
 # -----------------------------------------------
-def update(all=True, omni=False, leapsecs=False):
+def update(all=True, omni=False, leapsecs=False, callfromOMNI=False):
     """
     Download and update local database for omni, leapsecs etc
 
     Input:
     ======
-       - all (bol) : if True, update all of them
-       - omni (bol) : if True. update only onmi
-       - leapsecs (bol) : if True, update only leapseconds
+       - all (bool) : if True, update all of them
+       - omni (bool) : if True. update only onmi
+       - leapsecs (bool) : if True, update only leapseconds
 
 
      Example:
@@ -475,47 +475,55 @@ def update(all=True, omni=False, leapsecs=False):
     import urllib as u
     import os
     import zipfile
-    import spacepy.omni as om
+    oflag = 1
+    if not callfromOMNI:
+        oflag += 1
+        import spacepy.omni as om
     
-    dotfln = os.environ['HOME']+'/.spacepy'
-	#check directory exists, if not, create
-    if not os.path.isdir(dotfln):
-		os.mkdir(dotfln)
-		os.mkdir(dotfln+'/data')
-    datadir = dotfln+'/data'
-	# define location for getting leap seconds
-    leapsec_url ='ftp://maia.usno.navy.mil/ser7/tai-utc.dat'
-    leapsec_fname = dotfln+'/data/tai-utc.dat'
-	
-    # define location for getting omni
-    omni_url = 'ftp://virbo.org/QinDenton/hour/merged/latest/WGhour-latest.d.zip'
-    omni_fname_zip = dotfln+'/data/WGhour-latest.d.zip'
-    omni_fname_dat = dotfln+'/data/omnidata.pkl'
-	
-    if all == True:
-	    omni = True
-	    leapsecs = True
-	
-    if omni == True:
-	    # retrieve omni, unzip and save as table
-		print "Retrieving omni file ..."
-		u.urlretrieve(omni_url, omni_fname_zip)
-		fh_zip = zipfile.ZipFile(omni_fname_zip)
-		data = fh_zip.read(fh_zip.namelist()[0])
-		#dd = data.split('\n')
-		# save data as ascii file
-		fh = open(omni_fname_dat, 'w')
-		fh.writelines(data)
-		fh.flush()
-		fh.close
-		print "Now pickling (this will take a few minutes) ..."
-		om.pickleomni(fln=omni_fname_dat)
-		# delete left-overs
-		os.remove(omni_fname_zip)
-	
-    if leapsecs == True:
-        print "Retrieving leapseconds file ... "
-        u.urlretrieve(leapsec_url, leapsec_fname)
+    if oflag <= 1:
+        dotfln = os.environ['HOME']+'/.spacepy'
+        #check directory exists, if not, create
+        if not os.path.isdir(dotfln):
+            os.mkdir(dotfln)
+            os.mkdir(dotfln+'/data')
+        datadir = dotfln+'/data'
+        # define location for getting leap seconds
+        leapsec_url ='ftp://maia.usno.navy.mil/ser7/tai-utc.dat'
+        leapsec_fname = dotfln+'/data/tai-utc.dat'
+
+        # define location for getting omni
+        omni_url = 'ftp://virbo.org/QinDenton/hour/merged/latest/WGhour-latest.d.zip'
+        omni_fname_zip = dotfln+'/data/WGhour-latest.d.zip'
+        omni_fname_dat = dotfln+'/data/omnidata.pkl'
+
+        if all == True:
+            omni = True
+            leapsecs = True
+
+        if omni == True:
+            # retrieve omni, unzip and save as table
+            print "Retrieving omni file ..."
+            u.urlretrieve(omni_url, omni_fname_zip)
+            fh_zip = zipfile.ZipFile(omni_fname_zip)
+            data = fh_zip.read(fh_zip.namelist()[0])
+            #dd = data.split('\n')
+            # save data as ascii file
+            fh = open(omni_fname_dat, 'w')
+            fh.writelines(data)
+            fh.flush()
+            fh.close
+            print "Now pickling (this will take a few minutes) ..."
+            if not callfromOMNI:
+                om.pickleomni(fln=omni_fname_dat)
+                # delete left-overs
+                os.remove(omni_fname_zip)
+
+        if leapsecs == True:
+            print "Retrieving leapseconds file ... "
+            u.urlretrieve(leapsec_url, leapsec_fname)
+        
+    if callfromOMNI:
+        oflag += 1
 
     return datadir
